refactor(container): separate the different module
